requirements.txt
python-telegram-bot
numpy==1.12.0
psycopg2==2.6.2
